DbgSimpleSymbol can represent const data, and const statics are cached
Fixes a longstanding problem where DbgShell could not display the value of
static members that did not have an address (because they were constant
data).

This change also caches the symbol for static members, which can result in
significant speedups in certain cases: for instance, if you have a large
dictionary stored in a static T::x, then dump out a whole bunch of Ts, we
used to create a new DbgSymbol (and do Symbol Value Conversion for it) for
the "x" static member for every T, which was very wasteful and time-
consuming. Now we only do it once.

I've also stopped adding symbols representing static members as "children"
symbols. I don't think there was any good use for that.