Fix regression: console output encoding caused garbled output
In commit 07485ef29c, I added this as the first line of Main:

    Console.OutputEncoding = Encoding.Unicode;

With the comment:

    This allows "[Console]::WriteLine( [char] 0x2026 )" to work just as
    well as "[char] 0x2026".

But this caused output from other programs (that weren't emitting UTF16)
to be garbled (because the output was interpreted as UTF16 when captured
into a variable) (#95).

Hopefully setting the output encoding to UTF8 will work out better. At
least it lets both "[Console]::WriteLine( [char] 0x2026 )" and "git
--version" to work.

Addresses #95.